pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Manage system alises files."
authors = ["Andres <softwareapgm@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
logging = "^0.4.9.6"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
my-aliases = "app.command:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
